requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
